requirements.txt
streamlit
openai
fpdf2
pypdf
beautifulsoup4
requests
python-dotenv
regex
Pillow
boto3
botocore
python-docx
